feat:categorized transactions
Categorized transactions by each day

fixed minor bugs
