+++ app/src/main/java/com/diffy/broke/components/AddDialog.kt
@@ -34,8 +34,8 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import com.diffy.broke.Events
 import com.diffy.broke.States
-import com.diffy.broke.helpers.datePickerScreen
-import com.diffy.broke.helpers.dateInMillisToFormat
+import com.diffy.broke.utilcomponents.datePickerScreen
+import com.diffy.broke.utilcomponents.dateInMillisToFormat
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
