+++ app/src/main/java/com/diffy/broke/dataclasses/DataClasses.kt
@@ -2,6 +2,6 @@ package com.diffy.broke.dataclasses
 
 import com.diffy.broke.database.Transactions
 
-data class MinMaxDates(val minValue: Long, val maxValue: Long)
+data class MinMaxDates(val minValue: Long,val maxValue: Long)
 data class TransactionsInTimeperiod(val day: Long, val rangedTransactions: List<Transactions>)
-data class DayRange(val startTimeMillis: Long, val endTimeMillis: Long)
+data class SplittedDayRange(val startTimeMillis: Long, val endTimeMillis: Long)
