--- app/src/main/java/com/diffy/broke/helpers/HelperComponents.kt
@@ -1,150 +1,0 @@
-package com.diffy.broke.helpers
-
-import android.annotation.SuppressLint
-import android.util.Log
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.DatePicker
-import androidx.compose.material3.DatePickerDialog
-import androidx.compose.material3.DatePickerState
-import androidx.compose.material3.DateRangePicker
-import androidx.compose.material3.DateRangePickerState
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import java.text.SimpleDateFormat
-import java.util.Calendar
-import java.util.Locale
-
-@SuppressLint("SuspiciousIndentation")
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun datePickerScreen(
-    datePickerState: DatePickerState,
-    onShowDialogChange: (Boolean) -> Unit
-): Long {
-    var selectedDate: Long = datePickerState.selectedDateMillis!!
-        DatePickerDialog(
-            onDismissRequest = { onShowDialogChange(false) },
-            confirmButton = {
-                TextButton(onClick = {
-                    onShowDialogChange(false)
-                    selectedDate = datePickerState.selectedDateMillis!!
-                }) {
-                    Text("Ok")
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { onShowDialogChange(false) }) {
-                    Text("Cancel")
-                }
-            }
-        ) {
-            DatePicker(state = datePickerState)
-        }
-    return selectedDate
-}
-
-fun getTodayStartInMillis(): Long {
-    val calendar = Calendar.getInstance()
-    calendar.set(Calendar.HOUR_OF_DAY, 0)
-    calendar.set(Calendar.MINUTE, 0)
-    calendar.set(Calendar.SECOND, 0)
-    calendar.set(Calendar.MILLISECOND, 0)
-    return calendar.timeInMillis
-}
-
-fun getStartOfWeekInMillis(): Long {
-    val calendar = Calendar.getInstance()
-    calendar.set(Calendar.DAY_OF_WEEK, calendar.firstDayOfWeek)
-    calendar.set(Calendar.HOUR_OF_DAY, 0)
-    calendar.set(Calendar.MINUTE, 0)
-    calendar.set(Calendar.SECOND, 0)
-    calendar.set(Calendar.MILLISECOND, 0)
-    return calendar.timeInMillis
-}
-
-fun getStartOfMonthInMillis(): Long {
-    val calendar = Calendar.getInstance()
-    calendar.set(Calendar.DAY_OF_MONTH, 1)
-    calendar.set(Calendar.HOUR_OF_DAY, 0)
-    calendar.set(Calendar.MINUTE, 0)
-    calendar.set(Calendar.SECOND, 0)
-    calendar.set(Calendar.MILLISECOND, 0)
-    return calendar.timeInMillis
-}
-
-@Composable
-fun dateInMillisToFormat(dateMillis: Long?): String {
-    val currentDate = Calendar.getInstance()
-    val selectedDate = Calendar.getInstance().apply {
-        if (dateMillis != null) {
-            timeInMillis = dateMillis
-        }
-    }
-
-    return when {
-        currentDate.get(Calendar.YEAR) == selectedDate.get(Calendar.YEAR) &&
-                currentDate.get(Calendar.DAY_OF_YEAR) == selectedDate.get(Calendar.DAY_OF_YEAR) -> {
-            "Today"
-        }
-        else -> {
-            val dateFormat = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
-            dateFormat.format(selectedDate.time)
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun DateRangePickerScreen(
-    dateRangePickerState: DateRangePickerState,
-    onDismiss: () -> Unit,
-    setDateRangeOnApply: () -> Unit,
-) {
-    Dialog(
-        properties = DialogProperties(usePlatformDefaultWidth = false),
-        onDismissRequest = { onDismiss() }
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.background),
-            verticalArrangement = Arrangement.Top
-        ) {
-            DateRangePicker(state = dateRangePickerState, modifier = Modifier.weight(1f))
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(20.dp),
-                horizontalArrangement = Arrangement.End,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                TextButton(onClick = { onDismiss() }) {
-                    Text(text = "Cancel")
-                }
-                TextButton(onClick = {
-                    setDateRangeOnApply()
-                    println("Custom")
-                    Log.i("start", dateRangePickerState.selectedStartDateMillis!!.toString())
-                    Log.i("end", dateRangePickerState.selectedEndDateMillis!!.toString())
-                    onDismiss()
-                }) {
-                    Text(text = "Set")
-                }
-            }
-        }
-    }
-}
